@@ -3,7 +3,7 @@
   "name": "dh2o-atlantis",
   "author": "Digital H2O Team",
   "license": "MIT",
-  "version": "1.6.0",
+  "version": "1.7.0",
   "title": "DigitalH2O Style Guide",
   "description": "Style guide for front end apps",
   "bugs": {
@@ -36,7 +36,11 @@
     "babel-preset-stage-2": "^6.17.0",
     "cross-env": "^1.0.6",
     "css-loader": "^0.25.0",
-    "eslint": "^3.7.1",
+    "eslint": "^3.8.1",
+    "eslint-config-standard": "^6.2.1",
+    "eslint-plugin-promise": "^3.3.0",
+    "eslint-plugin-standard": "^2.0.1",
+    "eslint-plugin-html": "^1.3.0",
     "extract-text-webpack-plugin": "^2.0.0-beta.4",
     "file-loader": "^0.8.4",
     "js-beautify": "^1.6.4",
